added new thankyou to create s3 folder
created couple of new enhancements to the application Ui level and application functionality. s3 folder creation and upload document code.(It has an error to fix. and User index page has both codes, first one is with employee object and next one is related to the new object to carry the photo)